+++ keyboards/bioi/g60ble/rules.mk
@@ -9,7 +9,6 @@ MOUSEKEY_ENABLE = yes      # Mouse keys
 EXTRAKEY_ENABLE = yes      # Audio control and System control
 CONSOLE_ENABLE = no        # Console for debug
 COMMAND_ENABLE = yes       # Commands for debug and configuration
-NKRO_ENABLE = yes          # Enable N-Key Rollover
 BACKLIGHT_ENABLE = yes     # Enable keyboard backlight functionality
 RGBLIGHT_ENABLE = yes
 LTO_ENABLE = yes
